requirements/base.txt
python-dotenv==1.0.1
fastapi==0.110.2
pydantic==2.7.1
pydantic_core==2.18.2